requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-multipart==0.0.6
sentence-transformers==2.2.2
pdfplumber==0.7.7
python-docx==0.8.11
numpy==1.26.1
torch
